+++ servlib/bundling/BundlePayload.cc
@@ -252,14 +252,14 @@ BundlePayload::internal_write(const u_char* bp, size_t offset, size_t len)
         // that should cover it all
         else {
             PANIC("unexpected case in internal_write: "
-                  "data.length=%d offset=%d len=%d",
-                  data_.length(), offset, len);
+                  "data.length=%u offset=%u len=%u",
+                  (u_int)data_.length(), (u_int)offset, (u_int)len);
         }
 
         // sanity check
         ASSERTF(data_.length() >= offset + len,
-                "length=%d offset=%d len=%d",
-                data_.length(), offset, len);
+                "length=%u offset=%u len=%u",
+                (u_int)data_.length(), (u_int)offset, (u_int)len);
     }
     
     // check if we need to seek
@@ -339,8 +339,9 @@ BundlePayload::write_data(BundlePayload* src, size_t src_offset,
     oasys::ScopeLock l(lock_);
 
     log_debug("/bundle/payload",
-              "write_data: file=%s length_=%d src_offset=%d dst_offset=%d len %d", file_->path(),
-              length_, src_offset, dst_offset, len);
+              "write_data: file=%s length_=%u src_offset=%u dst_offset=%u len %u",
+              file_->path(),
+              (u_int)length_, (u_int)src_offset, (u_int)dst_offset, (u_int)len);
 
     ASSERT(length_       >= dst_offset + len);
     ASSERT(src->length() >= src_offset + len);
@@ -367,12 +368,12 @@ BundlePayload::read_data(size_t offset, size_t len, u_char* buf,
     oasys::ScopeLock l(lock_);
     
     ASSERTF(length_ >= (offset + len),
-            "length=%d offset=%d len=%d",
-            length_, offset, len);
+            "length=%u offset=%u len=%u",
+            (u_int)length_, (u_int)offset, (u_int)len);
 
     ASSERTF(rcvd_length_ >= (offset + len),
-            "rcvd_length=%d offset=%d len=%d",
-            rcvd_length_, offset, len);
+            "rcvd_length=%u offset=%u len=%u",
+            (u_int)rcvd_length_, (u_int)offset, (u_int)len);
     
     if (location_ == MEMORY) {
         return (u_char*)data_.data() + offset;
